Move InferredProxyContext to internal-api and InferredProxyPropagator to dd-trace-core
('components/context' should remain generic and not contain any specific propagator code)
