+++ dd-trace-core/src/main/java/datadog/trace/core/propagation/InferredProxyPropagator.java
@@ -1,7 +1,10 @@
-package datadog.context.propagation;
+package datadog.trace.core.propagation;
 
 import datadog.context.Context;
-import datadog.context.InferredProxyContext;
+import datadog.context.propagation.CarrierSetter;
+import datadog.context.propagation.CarrierVisitor;
+import datadog.context.propagation.Propagator;
+import datadog.trace.bootstrap.instrumentation.api.InferredProxyContext;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.BiConsumer;
